+++ packages/adminui-pub/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UBStoreManager`: add `description` to the shortcut attributes
 
 ### Changed
  - `UB.ux.Multifilter`: for composite attributes in `fieldList` without flag `simpleFilter: true`,